+++ mini-projects/day-04-function-02.py
@@ -1,0 +1,15 @@
+#Build a restaurant ordering system.
+#Create functions:
+#show_menu()
+#take_order()
+#show_bill()
+
+def show_menu(menu):
+    print("Menu :",menu)
+show_menu("chicken kebab")
+def take_order(order):
+    print("Order :",order)
+take_order("What kind")
+def show_bill(bill):
+    print("Bill :",bill)
+show_bill("Bill")